Fixed the old version of recreating course ids
