--- apps/cc-cmi5-player/src/app/components/MenuLayout.tsx
@@ -1,556 +1,0 @@
-/* eslint-disable react/jsx-no-useless-fragment */
-import { DRAWER_WIDTH } from './ContentLayout';
-import TabPanel from './TabPanel';
-import Drawer from '@mui/material/Drawer';
-import {
-  useCallback,
-  useEffect,
-  useLayoutEffect,
-  useRef,
-  useState,
-} from 'react';
-import { Panel, PanelGroup, PanelResizeHandle } from 'react-resizable-panels';
-import Stack from '@mui/material/Stack';
-import IconButton from '@mui/material/IconButton';
-import KeyboardArrowLeftIcon from '@mui/icons-material/KeyboardArrowLeft';
-import KeyboardArrowRightIcon from '@mui/icons-material/KeyboardArrowRight';
-import VerticalSplitIcon from '@mui/icons-material/VerticalSplit';
-import ReorderIcon from '@mui/icons-material/Reorder';
-
-/** Icons */
-import AccountCircleIcon from '@mui/icons-material/AccountCircle';
-import AssignmentIndIcon from '@mui/icons-material/AssignmentInd';
-import DarkModeIcon from '@mui/icons-material/DarkMode';
-import LightModeIcon from '@mui/icons-material/LightMode';
-import InfoOutlinedIcon from '@mui/icons-material/InfoOutlined';
-
-import { alpha, Box, Divider, Popover, Switch, Tooltip, Typography, useTheme } from '@mui/material';
-import { auJsonSel, classIdSel, studentIdSel } from '../redux/auReducer';
-import { activeTabSel } from '../redux/navigationReducer';
-import { useDispatch, useSelector } from 'react-redux';
-import { cmi5Instance } from '../session/cmi5';
-
-import { classChangeModalId } from './CourseModals';
-import RC5Player from './player/RC5Player';
-import ExitSlide from './ExitSlide';
-
-import ScenarioWrapper from './scenario/ScenarioWrapper';
-import { TeamScenarioContextProvider } from './team-consoles/TeamScenarioContext';
-import { useCMI5Session } from '../hooks/useCMI5Session';
-import {
-  ButtonInfoField,
-  ButtonInfoFormHeaderLayout,
-  ButtonMainUi,
-  config,
-  dividerColor,
-  setDividerColor,
-  setIconColor,
-  setModal,
-  setTheme,
-  maxSlideWidth$,
-  themeColor,
-} from '@rapid-cmi5/ui';
-import { CustomTheme } from '../styles/createPalette';
-
-/**
- * The main slide content.
- * Renders a layout with a collapsable menu on the left, and slide content on
- * the right.
- */
-export default function MenuLayout() {
-  const [isMenuDrawerOpen, setIsMenuDrawerOpen] = useState(true);
-  const [isSplitPanelShown, setIsSplitPanelShown] = useState(false);
-  const currentTheme: CustomTheme = useTheme();
-  const dispatch = useDispatch();
-  const classId = useSelector(classIdSel);
-  const studentId = useSelector(studentIdSel);
-  const auJson = useSelector(auJsonSel);
-  const activeTab = useSelector(activeTabSel);
-  const isExitSlide = activeTab === auJson?.slides?.length;
-  const themedDividerColor = useSelector(dividerColor);
-  const theColor = useSelector(themeColor);
-  const muiTheme = useTheme();
-  const { palette } = muiTheme;
-
-  const [password, setPassword] = useState('');
-  const [clearUserName, setCleanUserName] = useState('');
-  const [buildInfoAnchor, setBuildInfoAnchor] = useState<HTMLButtonElement | null>(null);
-  const slideRef = useRef<HTMLDivElement>(null);
-
-  const userName = cmi5Instance.getLaunchParameters().actor?.account?.name;
-  const regId = cmi5Instance.getLaunchParameters().registration;
-
-  const featureFlagChangeClassRoom = false;
-
-  useCMI5Session();
-
-  const getStrippedUserName = useCallback((inputName: string) => {
-    const pcteCharindex = inputName.indexOf('@pcte.mil');
-
-    if (pcteCharindex >= 0) {
-      const cleanName = inputName.substring(0, pcteCharindex);
-      setCleanUserName(cleanName);
-      setPassword(`${cleanName}!`);
-    } else {
-      setCleanUserName(inputName);
-      setPassword('');
-    }
-  }, []);
-
-  useEffect(() => {
-    if (userName) {
-      getStrippedUserName(userName);
-    }
-  }, [userName, getStrippedUserName]);
-
-  const handleMenuDrawerOpen = () => {
-    setIsMenuDrawerOpen(true);
-  };
-
-  const handleMenuDrawerClose = () => {
-    setIsMenuDrawerOpen(false);
-  };
-
-  const handleDividerOn = () => {
-    setIsSplitPanelShown(true);
-
-    // if the drawer is still open, close it
-    if (isMenuDrawerOpen) {
-      handleMenuDrawerClose();
-    }
-  };
-
-  const handleDividerOff = () => {
-    setIsSplitPanelShown(false);
-
-    // if the drawer is still closed, open it
-    if (!isMenuDrawerOpen) {
-      handleMenuDrawerOpen();
-    }
-  };
-
-  useEffect(() => {
-    if (currentTheme) {
-      const iconColor = currentTheme.button.iconColor;
-      const dividerColor = currentTheme.input.outlineColor;
-      dispatch(setIconColor(iconColor));
-      dispatch(setDividerColor(dividerColor || 'grey'));
-    }
-  }, [currentTheme, dispatch]);
-
-  /**
-   * Calculate left drawer width to support activity slides fullwidth css
-   * Observer ensures it gets updated when the left drawer is resized
-   */
-  useLayoutEffect(() => {
-    if (!slideRef.current) return;
-
-    const measure = () => {
-      if (slideRef.current) {
-        // Get the position relative to the viewport
-        const rect = slideRef.current.getBoundingClientRect();
-
-        //set width in css so we can tell activities to ignore content width settings
-        document.documentElement.style.setProperty(
-          '--panel-width',
-          `${rect.left}px`,
-        );
-
-        //store so we can constrain activities
-        maxSlideWidth$.value = rect.width;
-      }
-    };
-
-    const observer = new ResizeObserver(measure);
-    observer.observe(slideRef.current);
-
-    return () => observer.disconnect();
-
-    // Ensure the ref is attached to a DOM element and that element exists
-  }, []);
-
-  return (
-    <>
-      <Drawer
-        sx={{
-          width: isMenuDrawerOpen ? DRAWER_WIDTH : 0,
-          flexShrink: 0,
-          '& .MuiDrawer-paper': {
-            backgroundColor: 'background.default',
-            width: DRAWER_WIDTH,
-            boxSizing: 'border-box',
-            overflow: 'hidden',
-          },
-        }}
-        variant="persistent"
-        anchor="left"
-        open={isMenuDrawerOpen}
-      >
-        <div role="navigation" aria-label="Slides">
-          <TabPanel />
-        </div>
-      </Drawer>
-
-      <PanelGroup direction="horizontal">
-        <Panel defaultSize={45} minSize={5}>
-          <Box
-            ref={slideRef}
-            sx={{
-              //presentation background color
-              backgroundColor: palette.background.paper,
-
-              height: '100%',
-              width: '100%',
-              overflow: 'auto',
-            }}
-          >
-            <Stack
-              role="toolbar"
-              aria-label="Lesson Options"
-              direction="row"
-              sx={{
-                position: 'absolute',
-                zIndex: (theme) => theme.zIndex.appBar + 1,
-                backgroundColor: alpha(palette.background.paper, 0.9),
-                borderRadius: '6px',
-                width: '100%',
-              }}
-            >
-              <Tooltip
-                title={
-                  isMenuDrawerOpen ? 'Collapse Navigation' : 'Open Navigation'
-                }
-              >
-                <IconButton
-                  aria-label={
-                    isMenuDrawerOpen ? 'Collapse Navigation' : 'Open Navigation'
-                  }
-                  //tabIndex={-1}
-                  size={'small'}
-                  color="primary"
-                  onClick={() => {
-                    if (isMenuDrawerOpen) {
-                      handleMenuDrawerClose();
-                    } else {
-                      handleMenuDrawerOpen();
-                    }
-                  }}
-                >
-                  {isMenuDrawerOpen ? (
-                    <KeyboardArrowLeftIcon />
-                  ) : (
-                    <KeyboardArrowRightIcon />
-                  )}
-                </IconButton>
-              </Tooltip>
-
-              <Tooltip title="Turn Split Screen On">
-                <IconButton
-                  aria-label="Turn Split Screen On"
-                  //tabIndex={-1}
-                  color="primary"
-                  disabled={isSplitPanelShown}
-                  onClick={() => {
-                    if (isSplitPanelShown) {
-                      handleDividerOff();
-                    } else {
-                      handleDividerOn();
-                    }
-                  }}
-                >
-                  <VerticalSplitIcon />
-                </IconButton>
-              </Tooltip>
-              <Tooltip title="Turn Split Screen Off">
-                <IconButton
-                  aria-label="Turn Split Screen Off"
-                  color="primary"
-                  disabled={!isSplitPanelShown}
-                  onClick={() => {
-                    if (isSplitPanelShown) {
-                      handleDividerOff();
-                    } else {
-                      handleDividerOn();
-                    }
-                  }}
-                >
-                  <ReorderIcon />
-                </IconButton>
-              </Tooltip>
-              <Tooltip
-                title={
-                  theColor === 'dark'
-                    ? 'Switch to Light Mode'
-                    : 'Switch to Dark Mode'
-                }
-              >
-                <Switch
-                  checked={theColor === 'dark'}
-                  slotProps={{
-                    input: {
-                      'aria-label':
-                        theColor === 'dark'
-                          ? 'Switch to Light Mode'
-                          : 'Switch to Dark Mode',
-                      role: 'switch',
-                    },
-                  }}
-                  icon={
-                    <LightModeIcon
-                      name="Light Mode"
-                      color="primary"
-                      fontSize="small"
-                      sx={{
-                        backgroundColor: 'background.default',
-                        borderRadius: '12px',
-                        borderStyle: 1,
-                      }}
-                    />
-                  }
-                  checkedIcon={
-                    <DarkModeIcon
-                      name="Dark Mode"
-                      color="primary"
-                      fontSize="small"
-                      sx={{
-                        backgroundColor: 'background.default',
-                        borderRadius: '12px',
-                        borderStyle: 1,
-                      }}
-                    />
-                  }
-                  onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-                    if (event.target.checked) {
-                      dispatch(setTheme('dark'));
-                      return;
-                    }
-                    dispatch(setTheme('light'));
-                  }}
-                />
-              </Tooltip>
-              {featureFlagChangeClassRoom && (
-                <ButtonInfoField
-                  alertSxProps={{
-                    borderStyle: 'solid',
-                    borderWidth: '1px',
-                    borderColor: (theme: CustomTheme) =>
-                      `${theme.input.outlineColor}`,
-                    color: 'white',
-                  }}
-                  infoIcon={
-                    <AccountCircleIcon fontSize="inherit" color="primary" />
-                  }
-                  name="account-info-icon"
-                  message={
-                    <Stack
-                      direction="column"
-                      sx={{
-                        display: 'flex',
-                        marginLeft: '8px',
-                      }}
-                    >
-                      {studentId && (
-                        <Typography variant="caption">
-                          Student Id: {studentId}
-                        </Typography>
-                      )}
-                      <Typography variant="caption">
-                        Registration Id: {regId}
-                      </Typography>
-                      {classId && (
-                        <Typography variant="caption">
-                          Class Id: {classId}
-                        </Typography>
-                      )}
-                      <Typography variant="caption">
-                        User Name:
-                        {clearUserName}
-                      </Typography>
-                      {password && (
-                        <Typography variant="caption">
-                          Password:
-                          {password}
-                        </Typography>
-                      )}
-                      {classId && (
-                        <div>
-                          <ButtonMainUi
-                            startIcon={<AssignmentIndIcon />}
-                            onClick={() => {
-                              dispatch(
-                                setModal({
-                                  type: classChangeModalId,
-                                  id: '',
-                                  name: '',
-                                }),
-                              );
-                            }}
-                          >
-                            Change ClassRoom
-                          </ButtonMainUi>
-                        </div>
-                      )}
-                    </Stack>
-                  }
-                  props={{ sx: ButtonInfoFormHeaderLayout }}
-                  triggerOnClick={true}
-                />
-              )}
-
-              {auJson?.metadata && (
-                <>
-                  <Tooltip title="Build Information">
-                    <IconButton
-                      aria-label="Build Information"
-                      color="primary"
-                      size="small"
-                      onClick={(e) => setBuildInfoAnchor(e.currentTarget)}
-                    >
-                      <InfoOutlinedIcon fontSize="small" />
-                    </IconButton>
-                  </Tooltip>
-                  <Popover
-                    open={Boolean(buildInfoAnchor)}
-                    anchorEl={buildInfoAnchor}
-                    onClose={() => setBuildInfoAnchor(null)}
-                    anchorOrigin={{ vertical: 'bottom', horizontal: 'left' }}
-                    transformOrigin={{ vertical: 'top', horizontal: 'left' }}
-                    slotProps={{
-                      paper: {
-                        sx: {
-                          mt: 0.5,
-                          minWidth: 320,
-                          maxWidth: 480,
-                          borderRadius: 2,
-                          border: '1px solid',
-                          borderColor: 'divider',
-                        },
-                      },
-                    }}
-                  >
-                    <Box sx={{ px: 2.5, pt: 2, pb: 0.5 }}>
-                      <Typography
-                        variant="overline"
-                        sx={{ letterSpacing: 1.5, color: 'text.secondary', fontWeight: 600 }}
-                      >
-                        Build Information
-                      </Typography>
-                    </Box>
-                    <Divider />
-                    <Stack direction="column" spacing={1.5} sx={{ px: 2.5, py: 2 }}>
-                      {[
-                        { label: 'RC5 Version', value: auJson.metadata.rc5Version },
-                        { label: 'Branch', value: auJson.metadata.gitBranch },
-                        { label: 'Commit', value: auJson.metadata.commitHash },
-                        { label: 'Remote', value: auJson.metadata.remoteGitUrl },
-                        { label: 'Build Time', value: auJson.metadata.buildTime ? new Date(auJson.metadata.buildTime).toLocaleString() : undefined },
-                      ]
-                        .filter((row) => Boolean(row.value))
-                        .map(({ label, value }) => (
-                          <Stack key={label} direction="row" spacing={1} alignItems="flex-start">
-                            <Typography
-                              variant="caption"
-                              sx={{ color: 'text.secondary', minWidth: 80, pt: '1px', fontWeight: 600 }}
-                            >
-                              {label}
-                            </Typography>
-                            <Typography
-                              variant="caption"
-                              sx={{
-                                fontFamily: 'monospace',
-                                wordBreak: 'break-all',
-                                color: 'text.primary',
-                              }}
-                            >
-                              {value}
-                            </Typography>
-                          </Stack>
-                        ))}
-                    </Stack>
-                  </Popover>
-                </>
-              )}
-
-              {!featureFlagChangeClassRoom && (
-                <IconButton aria-label="Student Information" color="primary">
-                  <Tooltip
-                    arrow
-                    enterDelay={500}
-                    enterNextDelay={500}
-                    sx={{ maxWidth: '480px' }}
-                    title={
-                      <Stack
-                        direction="column"
-                        sx={{
-                          display: 'flex',
-                        }}
-                      >
-                        {studentId && (
-                          <Typography variant="caption">
-                            Student Id:{studentId}
-                          </Typography>
-                        )}
-                        <Typography variant="caption">
-                          Registration Id:{regId}
-                        </Typography>
-                        {classId && (
-                          <Typography variant="caption">
-                            Class Id:{classId}
-                          </Typography>
-                        )}
-                        <Typography variant="caption">
-                          User Name:
-                          {clearUserName}
-                        </Typography>
-                        {password && (
-                          <Typography variant="caption">
-                            Password:
-                            {password}
-                          </Typography>
-                        )}
-                      </Stack>
-                    }
-                    placement="bottom"
-                  >
-                    <AccountCircleIcon />
-                  </Tooltip>
-                </IconButton>
-              )}
-            </Stack>
-
-            <>
-              {isExitSlide && <ExitSlide />}
-              {!isExitSlide && !config.CMI5_SSO_ENABLED && (
-                <ScenarioWrapper>
-                  <RC5Player />
-                </ScenarioWrapper>
-              )}
-              {!isExitSlide && config.CMI5_SSO_ENABLED && (
-                <TeamScenarioContextProvider isEnabled={true}>
-                  <RC5Player />
-                </TeamScenarioContextProvider>
-                // REF debugging - keep for testing individual Scenario UI with deployed scenario
-                // <ScenarioWrapper>
-                //   <TeamScenarioContextProvider isEnabled={true}>
-                //     <RC5Player />
-                //   </TeamScenarioContextProvider>
-                // </ScenarioWrapper>
-              )}
-            </>
-          </Box>
-        </Panel>
-        {isSplitPanelShown && (
-          <>
-            <PanelResizeHandle
-              style={{
-                width: 4,
-                backgroundColor: themedDividerColor,
-              }}
-            />
-            <Panel></Panel>
-          </>
-        )}
-      </PanelGroup>
-    </>
-  );
-}
